User module Complete.Now user is added updated and deleted with all the error messages

+++ Flixer/views.py
@@ -38,10 +38,30 @@ def searchMovie(request):
 
 
 def deleteUser(request):
-    User.objects.filter(user_id=request.POST.get('user_id')).delete()
-    userPage(request)
+    if request.method == 'POST':
+        User.objects.filter(user_id=request.POST.get('user_id')).delete()
+        return userPage(request, 'deleted')
 
 
 def addUser(request):
     if request.method == 'POST':
         new_user = User()
+        new_user.user_id = request.POST.get('user_id')
+        new_user.name = request.POST.get('user_name')
+        new_user.password = request.POST.get('pass')
+        new_user.email = request.POST.get('emaill')
+        new_user.gender = request.POST.get('radio1')
+        new_user.save()
+        return userPage(request, 'yes')
+
+
+def editUser(request):
+    if request.method == 'POST':
+        new_user = User.objects.filter(user_id=request.POST.get('user_id1')).get()
+        new_user.name = request.POST.get('user_name1')
+        new_user.password = request.POST.get('pass1')
+        new_user.email = request.POST.get('emaill1')
+        new_user.gender = request.POST.get('rad')
+
+        new_user.save()
+        return userPage(request, 'updated')
